fix(live-engine): fractional crypto qty + SAS service auth post-migration
- live_trade_executions.requested_quantity Integer → Float (DB migration via
  migrate_qty_to_float.py). Strategies emitting 0.131 BTC were truncated to 0,
  causing all paper sessions on small capital (≤$10K) to record BUY 0 fills
  and never complete cycles. ETHUSDT (BUY 50, integer) worked; BTCUSDT 0.131
  did not.
- order_execution_service.py type hints int → float for quantity params.
- paper-scheduler.md: bind service user to ExchangeAccount id=12 after the
  user-table migration that brought local users (plzmlc + others) into mint.
  initial_capital lowered to \$1K — realistic minimum entry capital.
- SAS scripts: load Telegram bot creds from exchange_accounts (Fernet decrypt)
  instead of .env. New helper load_telegram_creds.py. wrapper + run_watchdog
  + run_live_monitor + run_meta_observer all updated. Aligns with
  feedback_credentials_in_db.md memory: .env holds system config only.

Verified on mint: \$1K paper sessions now produce real BUY/SELL cycles
(cmf_money_flow: 1 cycle within minutes of restart).

+++ backend/app/models/live_trading.py
@@ -122,7 +122,7 @@ class LiveTradeExecution(Base):
     signal_type = Column(String, nullable=False) # BUY / SELL
     signal_timestamp = Column(DateTime, nullable=False) # Chart Time (e.g. 09:00)
     theoretical_price = Column(Float, nullable=False) # Close Price of Signal Candle
-    requested_quantity = Column(Integer, default=0)
+    requested_quantity = Column(Float, default=0.0)  # Float to support fractional crypto qty (e.g. 0.131 BTC)
     
     # 2. Actual Execution
     exchange_order_no = Column(String, index=True, nullable=True)  # Kiwoom 주문번호 (체결 매칭용)
